requirements.txt
pandas>=1.3.0
python-docx>=0.8.11
flask>=2.0.0
python-dotenv>=0.19.0
flask-sqlalchemy>=2.5.0
Werkzeug>=2.0.0
email-validator>=1.1.3
gunicorn>=20.1.0
pytest>=6.2.5
black>=21.5b2
flake8>=3.9.0
cryptography>=3.4.7 